requirements.txt
# Invoice Control App - Dependencies
# Python 3.11+ required

streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pdf2image>=1.16.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0
